[dyn] synaptic projection updates
1. reorganize the projection structures;
2. rename previous reduced projections with intuitive names
3. add `brainpy.dyn.HalfProjDelta` and `brainpy.dyn.FullProjDelta`
