+++ brainpy/_src/dyn/neurons/lif.py
@@ -119,7 +119,7 @@ class IFLTC(GradNeuDyn):
       self.reset_state(self.mode)
 
   def derivative(self, V, t, I):
-    I = self.sum_inputs(V, init=I)
+    I = self.sum_current_inputs(V, init=I)
     return (-V + self.V_rest + self.R * I) / self.tau
 
   def reset_state(self, batch_size=None, **kwargs):
@@ -132,7 +132,7 @@ class IFLTC(GradNeuDyn):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    self.V.value = self.integral(self.V.value, t, x, dt)
+    self.V.value = self.integral(self.V.value, t, x, dt) + self.sum_delta_inputs()
 
     return self.V.value
 
@@ -146,7 +146,7 @@ class IF(IFLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -252,7 +252,7 @@ class LifLTC(GradNeuDyn):
       self.reset_state(self.mode)
 
   def derivative(self, V, t, I):
-    I = self.sum_inputs(V, init=I)
+    I = self.sum_current_inputs(V, init=I)
     return (-V + self.V_rest + self.R * I) / self.tau
 
   def reset_state(self, batch_size=None, **kwargs):
@@ -265,7 +265,7 @@ class LifLTC(GradNeuDyn):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    V = self.integral(self.V.value, t, x, dt)
+    V = self.integral(self.V.value, t, x, dt) + self.sum_delta_inputs()
 
     # spike, spiking time, and membrane potential reset
     if isinstance(self.mode, bm.TrainingMode):
@@ -337,7 +337,7 @@ class Lif(LifLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -464,7 +464,7 @@ class LifRefLTC(LifLTC):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    V = self.integral(self.V.value, t, x, dt)
+    V = self.integral(self.V.value, t, x, dt) + self.sum_delta_inputs()
 
     # refractory
     refractory = (t - self.t_last_spike) <= self.tau_ref
@@ -552,7 +552,7 @@ class LifRef(LifRefLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -723,7 +723,7 @@ class ExpIFLTC(GradNeuDyn):
       self.reset_state(self.mode)
 
   def derivative(self, V, t, I):
-    I = self.sum_inputs(V, init=I)
+    I = self.sum_current_inputs(V, init=I)
     exp_v = self.delta_T * bm.exp((V - self.V_T) / self.delta_T)
     dvdt = (- (V - self.V_rest) + exp_v + self.R * I) / self.tau
     return dvdt
@@ -738,7 +738,7 @@ class ExpIFLTC(GradNeuDyn):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    V = self.integral(self.V.value, t, x, dt)
+    V = self.integral(self.V.value, t, x, dt) + self.sum_delta_inputs()
 
     # spike, spiking time, and membrane potential reset
     if isinstance(self.mode, bm.TrainingMode):
@@ -880,7 +880,7 @@ class ExpIF(ExpIFLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -1076,7 +1076,7 @@ class ExpIFRefLTC(ExpIFLTC):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    V = self.integral(self.V.value, t, x, dt)
+    V = self.integral(self.V.value, t, x, dt) + self.sum_delta_inputs()
 
     # refractory
     refractory = (t - self.t_last_spike) <= self.tau_ref
@@ -1228,7 +1228,7 @@ class ExpIFRef(ExpIFRefLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -1400,7 +1400,7 @@ class AdExIFLTC(GradNeuDyn):
       self.reset_state(self.mode)
 
   def dV(self, V, t, w, I):
-    I = self.sum_inputs(V, init=I)
+    I = self.sum_current_inputs(V, init=I)
     exp = self.delta_T * bm.exp((V - self.V_T) / self.delta_T)
     dVdt = (- V + self.V_rest + exp - self.R * w + self.R * I) / self.tau
     return dVdt
@@ -1424,7 +1424,7 @@ class AdExIFLTC(GradNeuDyn):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    V, w = self.integral(self.V.value, self.w.value, t, x, dt)
+    V, w = self.integral(self.V.value, self.w.value, t, x, dt) + self.sum_delta_inputs()
 
     # spike, spiking time, and membrane potential reset
     if isinstance(self.mode, bm.TrainingMode):
@@ -1559,7 +1559,7 @@ class AdExIF(AdExIFLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -1756,7 +1756,7 @@ class AdExIFRefLTC(AdExIFLTC):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    V, w = self.integral(self.V.value, self.w.value, t, x, dt)
+    V, w = self.integral(self.V.value, self.w.value, t, x, dt) + self.sum_delta_inputs()
 
     # refractory
     refractory = (t - self.t_last_spike) <= self.tau_ref
@@ -1901,7 +1901,7 @@ class AdExIFRef(AdExIFRefLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -2040,7 +2040,7 @@ class QuaIFLTC(GradNeuDyn):
       self.reset_state(self.mode)
 
   def derivative(self, V, t, I):
-    I = self.sum_inputs(V, init=I)
+    I = self.sum_current_inputs(V, init=I)
     dVdt = (self.c * (V - self.V_rest) * (V - self.V_c) + self.R * I) / self.tau
     return dVdt
 
@@ -2054,7 +2054,7 @@ class QuaIFLTC(GradNeuDyn):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    V = self.integral(self.V.value, t, x, dt)
+    V = self.integral(self.V.value, t, x, dt) + self.sum_delta_inputs()
 
     # spike, spiking time, and membrane potential reset
     if isinstance(self.mode, bm.TrainingMode):
@@ -2166,7 +2166,7 @@ class QuaIF(QuaIFLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -2330,7 +2330,7 @@ class QuaIFRefLTC(QuaIFLTC):
     x = 0. if x is None else x
 
     # integrate membrane potential
-    V = self.integral(self.V.value, t, x, dt)
+    V = self.integral(self.V.value, t, x, dt) + self.sum_delta_inputs()
 
     # refractory
     refractory = (t - self.t_last_spike) <= self.tau_ref
@@ -2451,7 +2451,7 @@ class QuaIFRef(QuaIFRefLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -2609,7 +2609,7 @@ class AdQuaIFLTC(GradNeuDyn):
       self.reset_state(self.mode)
 
   def dV(self, V, t, w, I):
-    I = self.sum_inputs(V, init=I)
+    I = self.sum_current_inputs(V, init=I)
     dVdt = (self.c * (V - self.V_rest) * (V - self.V_c) - w + I) / self.tau
     return dVdt
 
@@ -2633,6 +2633,7 @@ class AdQuaIFLTC(GradNeuDyn):
 
     # integrate membrane potential
     V, w = self.integral(self.V.value, self.w.value, t, x, dt)
+    V = V + self.sum_delta_inputs()
 
     # spike, spiking time, and membrane potential reset
     if isinstance(self.mode, bm.TrainingMode):
@@ -2756,7 +2757,7 @@ class AdQuaIF(AdQuaIFLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -2939,6 +2940,7 @@ class AdQuaIFRefLTC(AdQuaIFLTC):
 
     # integrate membrane potential
     V, w = self.integral(self.V.value, self.w.value, t, x, dt)
+    V +=  self.sum_delta_inputs()
 
     # refractory
     refractory = (t - self.t_last_spike) <= self.tau_ref
@@ -3072,7 +3074,7 @@ class AdQuaIFRef(AdQuaIFRefLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -3279,7 +3281,7 @@ class GifLTC(GradNeuDyn):
     return self.a * (V - self.V_rest) - self.b * (V_th - self.V_th_inf)
 
   def dV(self, V, t, I1, I2, I):
-    I = self.sum_inputs(V, init=I)
+    I = self.sum_current_inputs(V, init=I)
     return (- (V - self.V_rest) + self.R * (I + I1 + I2)) / self.tau
 
   @property
@@ -3300,6 +3302,7 @@ class GifLTC(GradNeuDyn):
 
     # integrate membrane potential
     I1, I2, V_th, V = self.integral(self.I1.value, self.I2.value, self.V_th.value, self.V.value, t, x, dt)
+    V += self.sum_delta_inputs()
 
     # spike, spiking time, and membrane potential reset
     if isinstance(self.mode, bm.TrainingMode):
@@ -3452,7 +3455,7 @@ class Gif(GifLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -3680,6 +3683,7 @@ class GifRefLTC(GifLTC):
 
     # integrate membrane potential
     I1, I2, V_th, V = self.integral(self.I1.value, self.I2.value, self.V_th.value, self.V.value, t, x, dt)
+    V += self.sum_delta_inputs()
 
     # refractory
     refractory = (t - self.t_last_spike) <= self.tau_ref
@@ -3846,7 +3850,7 @@ class GifRef(GifRefLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -4012,7 +4016,7 @@ class IzhikevichLTC(GradNeuDyn):
       self.reset_state(self.mode)
 
   def dV(self, V, t, u, I):
-    I = self.sum_inputs(V, init=I)
+    I = self.sum_current_inputs(V, init=I)
     dVdt = self.p1 * V * V + self.p2 * V + self.p3 - u + I
     return dVdt
 
@@ -4040,6 +4044,7 @@ class IzhikevichLTC(GradNeuDyn):
 
     # integrate membrane potential
     V, u = self.integral(self.V.value, self.u.value, t, x, dt)
+    V += self.sum_delta_inputs()
 
     # spike, spiking time, and membrane potential reset
     if isinstance(self.mode, bm.TrainingMode):
@@ -4161,7 +4166,7 @@ class Izhikevich(IzhikevichLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
 
 
@@ -4351,6 +4356,7 @@ class IzhikevichRefLTC(IzhikevichLTC):
 
     # integrate membrane potential
     V, u = self.integral(self.V.value, self.u.value, t, x, dt)
+    V += self.sum_delta_inputs()
 
     # refractory
     refractory = (t - self.t_last_spike) <= self.tau_ref
@@ -4485,7 +4491,7 @@ class IzhikevichRef(IzhikevichRefLTC):
 
   def update(self, x=None):
     x = 0. if x is None else x
-    x = self.sum_inputs(self.V.value, init=x)
+    x = self.sum_current_inputs(self.V.value, init=x)
     return super().update(x)
