--- brainpy/_src/dyn/projections/__init__.py
@@ -1,5 +1,0 @@
-
-from .aligns import *
-from .conn import *
-from .others import *
-from .inputs import *
