+++ brainpy/_src/dyn/projections/tests/test_aligns.py
@@ -19,7 +19,7 @@ def test_ProjAlignPreMg1():
 
       prob = 80 / (4000 * scale)
 
-      self.E2I = bp.dyn.ProjAlignPreMg1(
+      self.E2I = bp.dyn.FullProjAlignPreSDMg(
         pre=self.E,
         syn=bp.dyn.Expon.desc(self.E.varshape, tau=5.),
         delay=delay,
@@ -27,7 +27,7 @@ def test_ProjAlignPreMg1():
         out=bp.dyn.COBA(E=0.),
         post=self.I,
       )
-      self.E2E = bp.dyn.ProjAlignPreMg1(
+      self.E2E = bp.dyn.FullProjAlignPreSDMg(
         pre=self.E,
         syn=bp.dyn.Expon.desc(self.E.varshape, tau=5.),
         delay=delay,
@@ -35,7 +35,7 @@ def test_ProjAlignPreMg1():
         out=bp.dyn.COBA(E=0.),
         post=self.E,
       )
-      self.I2E = bp.dyn.ProjAlignPreMg1(
+      self.I2E = bp.dyn.FullProjAlignPreSDMg(
         pre=self.I,
         syn=bp.dyn.Expon.desc(self.I.varshape, tau=10.),
         delay=delay,
@@ -43,7 +43,7 @@ def test_ProjAlignPreMg1():
         out=bp.dyn.COBA(E=-80.),
         post=self.E,
       )
-      self.I2I = bp.dyn.ProjAlignPreMg1(
+      self.I2I = bp.dyn.FullProjAlignPreSDMg(
         pre=self.I,
         syn=bp.dyn.Expon.desc(self.I.varshape, tau=10.),
         delay=delay,
@@ -90,7 +90,7 @@ def test_ProjAlignPostMg2():
 
       prob = 80 / (4000 * scale)
 
-      self.E2E = bp.dyn.ProjAlignPostMg2(
+      self.E2E = bp.dyn.FullProjAlignPostMg(
         pre=self.E,
         delay=delay,
         comm=bp.dnn.EventCSRLinear(bp.conn.FixedProb(prob, pre=self.E.num, post=self.E.num), 0.6),
@@ -98,7 +98,7 @@ def test_ProjAlignPostMg2():
         out=bp.dyn.COBA.desc(E=0.),
         post=self.E,
       )
-      self.E2I = bp.dyn.ProjAlignPostMg2(
+      self.E2I = bp.dyn.FullProjAlignPostMg(
         pre=self.E,
         delay=delay,
         comm=bp.dnn.EventCSRLinear(bp.conn.FixedProb(prob, pre=self.E.num, post=self.I.num), 0.6),
@@ -106,7 +106,7 @@ def test_ProjAlignPostMg2():
         out=bp.dyn.COBA.desc(E=0.),
         post=self.I,
       )
-      self.I2E = bp.dyn.ProjAlignPostMg2(
+      self.I2E = bp.dyn.FullProjAlignPostMg(
         pre=self.I,
         delay=delay,
         comm=bp.dnn.EventCSRLinear(bp.conn.FixedProb(prob, pre=self.I.num, post=self.E.num), 6.7),
@@ -114,7 +114,7 @@ def test_ProjAlignPostMg2():
         out=bp.dyn.COBA.desc(E=-80.),
         post=self.E,
       )
-      self.I2I = bp.dyn.ProjAlignPostMg2(
+      self.I2I = bp.dyn.FullProjAlignPostMg(
         pre=self.I,
         delay=delay,
         comm=bp.dnn.EventCSRLinear(bp.conn.FixedProb(prob, pre=self.I.num, post=self.I.num), 6.7),
@@ -163,14 +163,14 @@ def test_ProjAlignPost1():
       self.N = bp.dyn.LifRefLTC(num, V_rest=-60., V_th=-50., V_reset=-60., tau=20., tau_ref=5.,
                                 V_initializer=bp.init.Normal(-55., 2.))
       self.delay = bp.VarDelay(self.N.spike, entries={'I': None})
-      self.E = bp.dyn.ProjAlignPost1(comm=bp.dnn.EventJitFPHomoLinear(self.num_exc, num, prob=prob, weight=0.6),
-                                     syn=bp.dyn.Expon(size=num, tau=5.),
-                                     out=bp.dyn.COBA(E=0.),
-                                     post=self.N)
-      self.I = bp.dyn.ProjAlignPost1(comm=bp.dnn.EventJitFPHomoLinear(self.num_inh, num, prob=prob, weight=6.7),
-                                     syn=bp.dyn.Expon(size=num, tau=10.),
-                                     out=bp.dyn.COBA(E=-80.),
-                                     post=self.N)
+      self.E = bp.dyn.HalfProjAlignPost(comm=bp.dnn.EventJitFPHomoLinear(self.num_exc, num, prob=prob, weight=0.6),
+                                        syn=bp.dyn.Expon(size=num, tau=5.),
+                                        out=bp.dyn.COBA(E=0.),
+                                        post=self.N)
+      self.I = bp.dyn.HalfProjAlignPost(comm=bp.dnn.EventJitFPHomoLinear(self.num_inh, num, prob=prob, weight=6.7),
+                                        syn=bp.dyn.Expon(size=num, tau=10.),
+                                        out=bp.dyn.COBA(E=-80.),
+                                        post=self.N)
 
     def update(self, input):
       spk = self.delay.at('I')
@@ -198,30 +198,30 @@ def test_ProjAlignPost2():
                                 V_initializer=bp.init.Normal(-55., 2.))
       self.I = bp.dyn.LifRefLTC(ni, V_rest=-60., V_th=-50., V_reset=-60., tau=20., tau_ref=5.,
                                 V_initializer=bp.init.Normal(-55., 2.))
-      self.E2E = bp.dyn.ProjAlignPost2(pre=self.E,
-                                       delay=delay,
-                                       comm=bp.dnn.EventJitFPHomoLinear(ne, ne, prob=p, weight=0.6),
-                                       syn=bp.dyn.Expon(size=ne, tau=5.),
-                                       out=bp.dyn.COBA(E=0.),
-                                       post=self.E)
-      self.E2I = bp.dyn.ProjAlignPost2(pre=self.E,
-                                       delay=delay,
-                                       comm=bp.dnn.EventJitFPHomoLinear(ne, ni, prob=p, weight=0.6),
-                                       syn=bp.dyn.Expon(size=ni, tau=5.),
-                                       out=bp.dyn.COBA(E=0.),
-                                       post=self.I)
-      self.I2E = bp.dyn.ProjAlignPost2(pre=self.I,
-                                       delay=delay,
-                                       comm=bp.dnn.EventJitFPHomoLinear(ni, ne, prob=p, weight=6.7),
-                                       syn=bp.dyn.Expon(size=ne, tau=10.),
-                                       out=bp.dyn.COBA(E=-80.),
-                                       post=self.E)
-      self.I2I = bp.dyn.ProjAlignPost2(pre=self.I,
-                                       delay=delay,
-                                       comm=bp.dnn.EventJitFPHomoLinear(ni, ni, prob=p, weight=6.7),
-                                       syn=bp.dyn.Expon(size=ni, tau=10.),
-                                       out=bp.dyn.COBA(E=-80.),
-                                       post=self.I)
+      self.E2E = bp.dyn.FullProjAlignPost(pre=self.E,
+                                          delay=delay,
+                                          comm=bp.dnn.EventJitFPHomoLinear(ne, ne, prob=p, weight=0.6),
+                                          syn=bp.dyn.Expon(size=ne, tau=5.),
+                                          out=bp.dyn.COBA(E=0.),
+                                          post=self.E)
+      self.E2I = bp.dyn.FullProjAlignPost(pre=self.E,
+                                          delay=delay,
+                                          comm=bp.dnn.EventJitFPHomoLinear(ne, ni, prob=p, weight=0.6),
+                                          syn=bp.dyn.Expon(size=ni, tau=5.),
+                                          out=bp.dyn.COBA(E=0.),
+                                          post=self.I)
+      self.I2E = bp.dyn.FullProjAlignPost(pre=self.I,
+                                          delay=delay,
+                                          comm=bp.dnn.EventJitFPHomoLinear(ni, ne, prob=p, weight=6.7),
+                                          syn=bp.dyn.Expon(size=ne, tau=10.),
+                                          out=bp.dyn.COBA(E=-80.),
+                                          post=self.E)
+      self.I2I = bp.dyn.FullProjAlignPost(pre=self.I,
+                                          delay=delay,
+                                          comm=bp.dnn.EventJitFPHomoLinear(ni, ni, prob=p, weight=6.7),
+                                          syn=bp.dyn.Expon(size=ni, tau=10.),
+                                          out=bp.dyn.COBA(E=-80.),
+                                          post=self.I)
 
     def update(self, inp):
       self.E2E()
@@ -292,30 +292,30 @@ def test_ProjAlignPreMg1_v2():
                                 V_initializer=bp.init.Normal(-55., 2.))
       self.I = bp.dyn.LifRefLTC(ni, V_rest=-60., V_th=-50., V_reset=-60., tau=20., tau_ref=5.,
                                 V_initializer=bp.init.Normal(-55., 2.))
-      self.E2E = bp.dyn.ProjAlignPreMg1(pre=self.E,
-                                        syn=bp.dyn.Expon.desc(size=ne, tau=5.),
-                                        delay=delay,
-                                        comm=bp.dnn.JitFPHomoLinear(ne, ne, prob=p, weight=0.6),
-                                        out=bp.dyn.COBA(E=0.),
-                                        post=self.E)
-      self.E2I = bp.dyn.ProjAlignPreMg1(pre=self.E,
-                                        syn=bp.dyn.Expon.desc(size=ne, tau=5.),
-                                        delay=delay,
-                                        comm=bp.dnn.JitFPHomoLinear(ne, ni, prob=p, weight=0.6),
-                                        out=bp.dyn.COBA(E=0.),
-                                        post=self.I)
-      self.I2E = bp.dyn.ProjAlignPreMg1(pre=self.I,
-                                        syn=bp.dyn.Expon.desc(size=ni, tau=10.),
-                                        delay=delay,
-                                        comm=bp.dnn.JitFPHomoLinear(ni, ne, prob=p, weight=6.7),
-                                        out=bp.dyn.COBA(E=-80.),
-                                        post=self.E)
-      self.I2I = bp.dyn.ProjAlignPreMg1(pre=self.I,
-                                        syn=bp.dyn.Expon.desc(size=ni, tau=10.),
-                                        delay=delay,
-                                        comm=bp.dnn.JitFPHomoLinear(ni, ni, prob=p, weight=6.7),
-                                        out=bp.dyn.COBA(E=-80.),
-                                        post=self.I)
+      self.E2E = bp.dyn.FullProjAlignPreSDMg(pre=self.E,
+                                             syn=bp.dyn.Expon.desc(size=ne, tau=5.),
+                                             delay=delay,
+                                             comm=bp.dnn.JitFPHomoLinear(ne, ne, prob=p, weight=0.6),
+                                             out=bp.dyn.COBA(E=0.),
+                                             post=self.E)
+      self.E2I = bp.dyn.FullProjAlignPreSDMg(pre=self.E,
+                                             syn=bp.dyn.Expon.desc(size=ne, tau=5.),
+                                             delay=delay,
+                                             comm=bp.dnn.JitFPHomoLinear(ne, ni, prob=p, weight=0.6),
+                                             out=bp.dyn.COBA(E=0.),
+                                             post=self.I)
+      self.I2E = bp.dyn.FullProjAlignPreSDMg(pre=self.I,
+                                             syn=bp.dyn.Expon.desc(size=ni, tau=10.),
+                                             delay=delay,
+                                             comm=bp.dnn.JitFPHomoLinear(ni, ne, prob=p, weight=6.7),
+                                             out=bp.dyn.COBA(E=-80.),
+                                             post=self.E)
+      self.I2I = bp.dyn.FullProjAlignPreSDMg(pre=self.I,
+                                             syn=bp.dyn.Expon.desc(size=ni, tau=10.),
+                                             delay=delay,
+                                             comm=bp.dnn.JitFPHomoLinear(ni, ni, prob=p, weight=6.7),
+                                             out=bp.dyn.COBA(E=-80.),
+                                             post=self.I)
 
     def update(self, inp):
       self.E2E()
@@ -350,30 +350,30 @@ def test_ProjAlignPreMg2():
                                 V_initializer=bp.init.Normal(-55., 2.))
       self.I = bp.dyn.LifRefLTC(ni, V_rest=-60., V_th=-50., V_reset=-60., tau=20., tau_ref=5.,
                                 V_initializer=bp.init.Normal(-55., 2.))
-      self.E2E = bp.dyn.ProjAlignPreMg2(pre=self.E,
-                                        delay=delay,
-                                        syn=bp.dyn.Expon.desc(size=ne, tau=5.),
-                                        comm=bp.dnn.JitFPHomoLinear(ne, ne, prob=p, weight=0.6),
-                                        out=bp.dyn.COBA(E=0.),
-                                        post=self.E)
-      self.E2I = bp.dyn.ProjAlignPreMg2(pre=self.E,
-                                        delay=delay,
-                                        syn=bp.dyn.Expon.desc(size=ne, tau=5.),
-                                        comm=bp.dnn.JitFPHomoLinear(ne, ni, prob=p, weight=0.6),
-                                        out=bp.dyn.COBA(E=0.),
-                                        post=self.I)
-      self.I2E = bp.dyn.ProjAlignPreMg2(pre=self.I,
-                                        delay=delay,
-                                        syn=bp.dyn.Expon.desc(size=ni, tau=10.),
-                                        comm=bp.dnn.JitFPHomoLinear(ni, ne, prob=p, weight=6.7),
-                                        out=bp.dyn.COBA(E=-80.),
-                                        post=self.E)
-      self.I2I = bp.dyn.ProjAlignPreMg2(pre=self.I,
-                                        delay=delay,
-                                        syn=bp.dyn.Expon.desc(size=ni, tau=10.),
-                                        comm=bp.dnn.JitFPHomoLinear(ni, ni, prob=p, weight=6.7),
-                                        out=bp.dyn.COBA(E=-80.),
-                                        post=self.I)
+      self.E2E = bp.dyn.FullProjAlignPreDSMg(pre=self.E,
+                                             delay=delay,
+                                             syn=bp.dyn.Expon.desc(size=ne, tau=5.),
+                                             comm=bp.dnn.JitFPHomoLinear(ne, ne, prob=p, weight=0.6),
+                                             out=bp.dyn.COBA(E=0.),
+                                             post=self.E)
+      self.E2I = bp.dyn.FullProjAlignPreDSMg(pre=self.E,
+                                             delay=delay,
+                                             syn=bp.dyn.Expon.desc(size=ne, tau=5.),
+                                             comm=bp.dnn.JitFPHomoLinear(ne, ni, prob=p, weight=0.6),
+                                             out=bp.dyn.COBA(E=0.),
+                                             post=self.I)
+      self.I2E = bp.dyn.FullProjAlignPreDSMg(pre=self.I,
+                                             delay=delay,
+                                             syn=bp.dyn.Expon.desc(size=ni, tau=10.),
+                                             comm=bp.dnn.JitFPHomoLinear(ni, ne, prob=p, weight=6.7),
+                                             out=bp.dyn.COBA(E=-80.),
+                                             post=self.E)
+      self.I2I = bp.dyn.FullProjAlignPreDSMg(pre=self.I,
+                                             delay=delay,
+                                             syn=bp.dyn.Expon.desc(size=ni, tau=10.),
+                                             comm=bp.dnn.JitFPHomoLinear(ni, ni, prob=p, weight=6.7),
+                                             out=bp.dyn.COBA(E=-80.),
+                                             post=self.I)
 
     def update(self, inp):
       self.E2E()
