+++ brainpy/dyn/projections.py
@@ -1,24 +1,24 @@
-
-from brainpy._src.dyn.projections.aligns import (
-  VanillaProj,
-  ProjAlignPostMg1,
-  ProjAlignPostMg2,
-  ProjAlignPost1,
-  ProjAlignPost2,
-  ProjAlignPreMg1,
-  ProjAlignPreMg2,
-  ProjAlignPre1,
-  ProjAlignPre2,
+from brainpy._src.dyn.projections.vanilla import VanillaProj
+from brainpy._src.dyn.projections.delta import (
+  HalfProjDelta,
+  FullProjDelta,
 )
-
+from brainpy._src.dyn.projections.align_post import (
+  HalfProjAlignPostMg,
+  FullProjAlignPostMg,
+  HalfProjAlignPost,
+  FullProjAlignPost,
+)
+from brainpy._src.dyn.projections.align_pre import (
+  FullProjAlignPreSDMg,
+  FullProjAlignPreDSMg,
+  FullProjAlignPreSD,
+  FullProjAlignPreDS,
+)
 from brainpy._src.dyn.projections.conn import (
   SynConn as SynConn,
 )
-
-from brainpy._src.dyn.projections.others import (
-  PoissonInput as PoissonInput,
-)
-
 from brainpy._src.dyn.projections.inputs import (
   InputVar,
+  PoissonInput,
 )
