+++ brainpy/dyn/synapses.py
@@ -1,6 +1,5 @@
 
 from brainpy._src.dyn.synapses.abstract_models import (
-  Delta,
   Expon,
   Alpha,
   DualExpon,
